requirements.txt
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
distlib==0.3.6
dulwich==0.21.3
filelock==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.1.1
poetry==1.4.0
poetry-core==1.5.2
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
tomlkit==0.11.6
trove-classifiers==2023.3.9
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0

mysql-connector-python~=8.0.32